requirements.txt
openai~=1.73.0
parea-ai~=0.2.220
numpy~=1.24.3
venn~=0.1.3
requests~=2.30.0
loguru~=0.5.3
pandas~=2.0.2
pyperclip~=1.8.2
python-dotenv~=1.0.1